+++ crates/scanner/src/interface/runner.rs
@@ -2,7 +2,9 @@ use crate::core::scan::{ScanConfig, ScanFilter, ScanMode};
 use crate::controller::notify::Notify;
 use crate::app::scan_task::ScanTask;
 use parser::core::encounter::MatchNature;
+#[allow(unused)]
 use parser::core::seat::SeatComposition;
+use tokio::sync::watch;
 
 pub struct ScannerHandle {
     abort_handle: tokio::task::AbortHandle,
@@ -18,26 +20,20 @@ impl ScannerHandle {
                 parser::core::club::ClubType::StadeRochelais,
                 "https://billetterie.staderochelais.com/fr".to_string(),
             ),
-            MatchNature::Basketball,
+            MatchNature::Rugby,
             Some(ScanFilter {
                 price_threshold: None,
                 date_range: None,
-                position: Some(SeatComposition {
-                    category: "".to_string(),
-                    bloc: "".to_string(),
-                    row: "".to_string(),
-                    seat_number: 0,
-                }),
+                position: None,
                 side_by_side: None,
-                // match_title: Some("STADE ROCHELAIS BASKET / ROUEN".to_string()),
-                match_title: Some("STADE ROCHELAIS / UNION BORDEAUX BÈGLES".to_string()),
+                match_title: None,
                 is_preview: Some(true),
             }),
         )
     }
     
-    pub fn start(config: ScanConfig, notifier: impl Notify) -> Self {
-        let task = ScanTask::new(config, notifier);
+    pub fn start(config_rx: watch::Receiver<ScanConfig>, notifier: impl Notify) -> Self {
+        let task = ScanTask::new(config_rx, notifier);
         let handle = tokio::spawn(async move { task.run().await });
         Self { abort_handle: handle.abort_handle() }
     }